requirements.txt
```
google-generativeai
streamlit
plotly
pandas
numpy
scikit-learn
faker